requirements.txt
Django==1.11.7
django-extensions==1.9.7
mysqlclient==1.3.12
MySQL-python==1.2.5
model-mommy>=1.4.0
django-multiselectfield==0.1.8
